\name{IlluminaHumanMethylation27kanno.ilmn12.hg19}
\alias{IlluminaHumanMethylation27kanno.ilmn12.hg19}
\docType{data}
\title{
  Annotation data for the \sQuote{IlluminaHumanMethylation27k} micro array.
}
\description{
 This package is based on the file
  \sQuote{HumanMethylation27_v-1-2.csv} from the zip archive
  from \url{www.illumina.com}.

  Additional SNP annotation is generated by the authors (described below).

  A script for creating the data object in this package is contained in
  \code{scripts/manifest.R}.
}
\usage{data(IlluminaHumanMethylation27kanno.ilmn12.hg19)}
\format{
  An object of class \code{IlluminaMethylationAnnotation}.
}
\details{
  The following changes/ modifications / addition has been made to the
  source material.

  The creation of this object based on the Illumina annotation and
  additional SNP information (see below) is contained in the
  \code{createAnno.R} script in the \code{scripts} directory.

  In addition to the SNP information provided by Illumina, we have added
  independent information on the overlap of the 27k with various
  versions of dbSNP.  The overlap is based on the
  mappings of the 27k loci to the hg19 genome provided by Illumina.  As
  dbSNP we have used the \sQuote{Common} table from UCSC
  (ie. \sQuote{snp138Common}).  This track contains variants from dbSNP
  which have a minor allele frequency (MAF) of greater than 1 percent
  (specifically, this requires dbSNP to actually contain MAF
  information).  Furthermore, we only kept variants marked as
  \sQuote{single} (ie. standard single nucleotide changes, but not
  insertions or deletions).  Scripts for retrieving the UCSC dbSNP table
  and doing the overlap are contained in the \code{scripts} directory.
  The variants are described in 6 different columns. \code{Probe_rs}
  tells us the RS number (SNP ID number) for a SNP overlapping the
  probe, and \code{Probe_maf} is the minor allelle frequency for the SNP
  (in case multiple SNPs overlap, only one is recorded).  Similarly,
  \code{CpG_rs} describe SNPs overlapping the CpG site and \code{SBE_rs}
  the single base extension of the measured methylation loci.
}

\source{
  See description.
}

\examples{
  data(IlluminaHumanMethylation27kanno.ilmn12.hg19)
}
\keyword{datasets}
